Iss #102: Fix bugs related to horiz. interp. degree
These bugs were in gridstatevector_mod. The hInterpolateDegree_opt
argument in calls to gsv_allocate were not being set for temporary
stateVectors created for use with gsv_hInterpolate. Therefore, the
default "linear" interpolation degree was always being used.

To ensure this problem does not happen again, the default values
for interpolation degree were removed such that if a user does
not explicitly set the interpolation degree in a statevector
given to gsv_hInterpolate, the code will abort with a useful
message in the listing.