Issue #104: create a new variable 'ensFileExtLengthStr'
Instead of reusing 'memberIndexStr', which may be misleading, we create
a new variable 'ensFileExtLengthStr' which contains a string with the
integer 'ensFileExtLength'.

We also remove the dynamic allocation of variables 'memberIndexStr' and
'ensFileExtLengthStr' which introduces non-desirable complexity.  We ask
for a string a 10 characters which is enough until we reach 10^10
members.

We also extend the variable 'ensNumber' in file 'fln_ensFileName' to 10
from 4.  The previous value, 4, was enough for not more than 10000
members.