#119 Changes to the advection on thermo levels
The advection of variables on thermodynamic levels was based on trajectories obtained from interpolated winds onto the the thermo levels. Insted, like in GEM, the positions found on the momentum levels are now interpolated onto the thermo levels. This change removes the previoustly observed inconsistency in the results on thermo levels found when using the "implicit" advection approach proposed by Meteo-France compare to "explicit" approach where the advection is only performed on momentum levels.