Issue #123: Fix bug related to interpolation of rotated winds
The complementary wind component was not being communicated
in transposeLatLonToVarsLevs in gridStateVector_mod. This
caused incorrect interpolated wind increments when requesting
to write the full analysis in the same execution as doing
the 3DVar or EnVar analysis.

For testing, the "gbc" system test was modified so that it
outputs the interpolated increment (rehm) and full analysis
(anlm) using "CUBIC" horizontal interpolation. This causes the
system test to fail, due to these additional output files.
They will need to be added to the reference results.