Issue #121: Add a contributing guide
A contributing guide has been created by only copying the one from the
project 'derivate':
    https://gitlab.science.gc.ca/atmospheric-data-assimilation/derivate/raw/6d4fd691f8b0051d8f0dfe309cca138d407c3cc1/CONTRIBUTING.md

This contributing guide, written by Josee Morneau (@jmo000), follows
modern software development guidelines.  GitLab proposes to read this
contributing guide automatically when the user wants to create an
issue.  That will help users to develop MIDAS and uniformize the style
of coding of everyone.

It will be adapted to MIDAS in the next commits.