Issue #125: some fixes to sphinx-fortran and added all programs
Fixed some major problems in sphinx-fortran related to
the description of subroutine arguments that used other
arguments to explicitly define its dimensions. Also fixed
a problem with module variables that were initialized in
the declaration statement (before these variables became
"optional").

Also, added all remaining MIDAS programs to the documentation.
Noticed that the subroutines "contained" in a main program are
not included in the documentation. This will have to be fixed
later.