Issue #121: propose to work on a new branch to 'pull --rebase'.
It will be safer to suggest people to make a copy and rebase the
copy (i.e. _v1), instead of rebasing the original branch which may
already be pushed?  This way they don't have to use "git push -f" and
just in case someone else is doing any work related to the original
pushed branch, it won't cause problems or confusion. This would also be
consistent with what is described on the wiki.