Issue #121: make commit message writing easier
In the previous commit message template, we add to erase nearly all the
first line before starting to write the message.

I choose to put the same information in the body of the template.  So the
user can read it while composing the title and the rest of the commit
message.

We also suggest the user reread himself.