Issue #128: Hessian IO bug found and fixed
The bug in the Hessian IO was due to incorrect specification of the size of the controlvector in LAM mode in bmatrix_mod.ftn90. Because no LAM option were implemented for bmatTypeList = CHM, the size of cvdim for that subvector was not set to zero (the previous value set in lbhi_Setup or ben_Setup was still assigned to cvDim) leading to an unwanted increase in the size of the control vector.

Fortunately, this bug had no impact on the LAM results but bmat_expand and bmat_preduce were messed up, leading to wrong writing and reading of the Hessian matrix.