Issue #129: Bugfix to statetocolumn_mod.f90
The bug was occuring in subroutine gd2mvoad in the extremely rare case where a processor had some observations before the load balancing but none after. In that case, the number of vertical level in the column structure (variable numLev) was uncorrectly set to the default value (i.e. 1) when no observation are present (because the info was based on the number of observation handled after the balancing step, which is correct in the forward operator but not in the adjoint version).

The appropriate change was made to insure that the correct value for numLev is always set in gd2mvoad.