UnitTest/check: remove nodelogger message when a file validates
When checking the results of a run, we were issuing a message for each
file to indicate if it validates or not.  Since there are many many
files to check, there are a lot of messages issued and this is slowing
xflow at a point that it can hardly be used.  Since the message that a
file validates is not very helpful in most of the cases, we are more
interested in the ones that do not validate, we do not send a message
when a file validates.  We send a message only if the file does not
validate.