Issue #126: Move routines 'min_{get,write}Increment' in 'increment_mod'
To group together all the routines related to analysis increments, we
move the two routine 'min_getIncrement' and 'min_writeIncrement' which
were in 'minimization_mod' into 'increment_mod'.  By the way, we changed
their prefixes from 'min_' to 'inc_'.

The only non-backward compatible change is the namelist variable
'CETIKINC' which was in 'NAMMIN'.  This variable was removed and
replaced by the variable 'etiket_rebm' in 'NAMINC'.