Issue #126: Move 'bmat_finalize' from 'var.f90' to 'minimization_mod.f90'
The computation of the high resolution ('rehm') and analysis ('anlm')
from trials ('trlm') and low resolution increments ('rehm') takes a lot
of memory.  So we want to free some as early as possible.  Much of the
memory is contained in the B-matrix representation.  So, we moved the
'bmat_finalize' call from 'var.f90' into routine 'min_minimize'.  We
also had to move 'osd_ObsSpaceDiag' since it is using the B-matrix.