Issue #126: The minimize module only computes the optimal control vector
With this commit, we want the 'minimization_mod' module to focus on its
main task which is to computes the optimal control vector with respect
to the cost function.  Other manipulations have been transferred into
the 'var' program.

Then, the computation of the increment, its writing to file, the
computation of high resolution increment and to full analysis are all
now done in the 'var' program.  To do so, we had to move the variable
'WRITEANALYSIS' from the 'NAMMIN' namelist to 'NAMCT0'.  All the tests
have been modified accordingly.