Issue #126: Remove 'use ... , only:' statements
We remove the 'use ... , only:' statements for the following reasons:
  * we have not been using it until now,
  * I don't find it very useful since if you don't use it, then
    everything is accessible, also
  * if you do specify things after "only" that you don't use, it won't
    complain, so it can be misleading.

In the end, to use the "only" correctly and throughout the entire code
would be a lot of manual checking after each modification to make sure
everything is up to date. Since it is not enforced, people just won't do
this, so it won't be uniformly used. And even if we did use it, we don't
see how it helps make the code easier to understand.

Our use of prefixes for all public entities is a much better way of
knowing in which module something is defined and this does not require
any work on the side of a module user to keep things up to date.