Issue #133: Copy files to ramdisk in fortran
Instead of relying on the script that launches the MIDAS
executable, the fortran code itself can now copy files
from disk to ramdisk. Therefore, the system test script
"var.sh" was simplified significantly to test this new
functionality in all tests that use the ramdisk.

Also added a new module containing the interfaces for
all "clib" subroutines supplied by rmnlib. This avoids
the need to include the clib "include" file in each module
that uses one of the clib subroutines, which would
require switching the file extension back to ftn90
from f90. The new module is: clib_interfaces_mod.ftn90.