Issue #133: Improve efficiency of fileCopy
The efficiency is improved by avoiding the copy of
files to ramdisk when it is only needed for setting
up the horizontal and vertical coordinate objects
from information in a standard file. This is because
all MPI tasks were trying to copy the same file,
which apparently caused a major slow down.