Issue #133: Impose the SSM domains in 'run' task
Instead of specifying for each test which SSM domain to load, we impose
a default list.  This depends on the platform.  For example, on PPP, we
must load explicitely OpenMPI but not on XC40.  So we can remove the
variable 'UnitTest_run_ssm' in each system test configuration.  We put a
default value which point to 'rpn/utils' since this is loaded for each
test.  We need 'r.run_in_parallel'.