Issue #134: Modify small number in vertical interp for tovs
This value was found to be not small enough in one
run, just by chance, which caused an abort. A better
solution may be available from Yves Rochon.

Also, initialize module variable dzmax in gps_mod to avoid
doing something with an uninitialized variable in cases
that setup routine is not called.