Issue #134: use improved version of windRotation_mod
This version uses simplified calculations that are a
little easier to follow. It also uses a derived type
to store the configuration, that will allow the multiple
configurations needed for the new unified interpolation.

Also, the setup to the windRotation_mod object is called
from closer to where it is used and not from the high-level
main programs.

The results for the biFourier system test were slightly
changed for rebm only and therefore were updated.