Issue #136: Move observations copy to './obs' in main task.
By doing so, we are sure that all MPI tasks are starting almost at the
same time.  When a big delay occurs between the start of each MPI task,
we obtain sometimes an abort with a 'Alarm clock' error message.  So the
strategy here is to do the less possible work in the script run in
MPI (which is 'var.sh') to minimize the delay between the start of the
MPI binary on each MPI tile.  Here, this reduces to create the ramdisk
directory and initialize the 'MIDAS_RAMDISKDIR' variable.