Issue #134: Centralize unit conversion and convert to Kelvin
Until now we have NOT been systematically converting the TT
field from Celcius to Kelvin when reading standard files. This
is now done, which requires additional information on whether
a file contains a "fullField" or not when reading or writing.
A new argument "containsFullField_opt" was added to several
routines to supply this information.

In addition, the conversion from "fileUnits" to "stateUnits" has
now been centralized into a new subroutine gsv_fileUnitsToStateUnits.
This will be needed when the new interpolation is introduced,
since several routines that read files will be added that all
need to perform unit conversion.

These change result in minor changes to the results of many
of the system tests due to slight changes when reading the
ensemble (due to conversion from C to K) or when reading
the background state.