Issue #136: Convert 'var.sh' to '/bin/bash'
Le shell '/bin/ksh' is sometimes very very slow and induces enornous
different timings between the different MPI tiles.  So, some MPI tile
are at the MPI part of the script as other are still in the script.
When this delay is too much, around 1 minute, the MPI program aborts
with an 'Alarm call' or 'Alarm clock'.

Using '/bin/bash', we are not seeing those big timing differences.