Issue #124: Simplifying the code.
Handling of GZ/P variables are simplified.

It is decided to process GZ/P separately from other variables to
avoid introducing too many changes to the code.

In gd2mv/commLatLon, select cases are removed by introducing
getFieldsWithHalo function. The modifications are reduced to minimize
the changes for the adjoint of the interpolator.