Adapt 'UnitTest/run' for inputs with 'obsfiles*'.
Some code was missing to make the tasks 'UnitTest/run' and 'var.sh' work
with 'obsfiles*' directories (prefix 'brp', 'obs' or 'sql').  We had to
modify the results of 'obsSelect/thin_AL' not for the content of files
but only the names: 'sqlfiles_al.updated/obsal_0001_0001' became
'obsfiles_al.updated/obsal_0001_0001'.