Issue #144: Impose a minimum value for 'HU' in 'inn_setupBackgroundColumnsAnl'
This minimum was imposed before #67 (and !111) but was removed.  This
minimum must still be there.  If not we obtain some negaitve values for
'HU' which, when we take the 'log', the program aborts.

The change in the minimum humidity affects extremely weakly the results
for 'var' program.  The 'var/EnVar/rdps' was not affected at all!

The test 'var/EnVar/gdps' was updated to use trial generated with GEM5
at 15km resolution.  Without the fix to the code, this code is aborting.
It was verified that, with the previous inputs for 'gdps' test', the
results are only impacted by a very small margin.