Issue #137: Using inputs with no corrupted files for 'var/bgck/conv'
The version 2 of the inputs were taken from a date on which the previous
versions aborted.  The problem was in part due to unrealistic values of
lat-lon that come from a corrupted BURP file.  Using this new set of
files as input for the test was giving irreproducible results.  Once per
5 executions, the results were different.  We could not traceback the
problem into the code.

So we decided to update the inputs and results with new files coming
from the operational G2 on 2018100906.  This set of files gives
reproducible results.