Issue #149: Remove last subroutines and automatic src_files
The last stand-alone subroutines/functions that were outside
of modules are now in the new modules `backgroundCheck_mod`
and `compute_HBHT_mod`. Also the last old-style fortran file
quasinewton_mod.f was converted to quasinewton_mod.f90.

A script was introduced that automatically generates the
src_files file for each program: `make_src_files.sh`. It
also creates a file `full_tree_${program}.dat` that contains
the entire modules dependency tree of each program (which
may or may not be useful in some situations). These are
currently not saved as part of the git repository. To faciliate
this new script, the additional configuration information that
was previously in each src_files file was put into new files:
`compile_setup_${program}.sh`.