Issue #80: Several final adjustments
Now gdUV is always available, even when both
wind components are available on the same mpi
task (in which case it points to the existing
complementary component and does not take any
extra memory).

To save memory when doing EnVar, the trials
are now read and interpolated to the column
before the B matrix setup. Because of this
I went back to reading all variables at once.
This is closer to what will be needed for
computing GZ and Pressure on the grid.

Updated references results for recenterLAM.

Some cleaning up of TMG timing blocks.

Switched to simpler interpolation in Yin-Yan
overlap to be more similar with ezsint.

Corrected the determination of "rotated" in
hco_setupFromFile for the Z grid. This changed
(corrected) the result for obs on equator when
background state is unrotated Z grid.

Corrected var.f90 which was missing calls to
conventional background check (lost in a rebase)!

Added a default reference_update path for all
system tests, for convenience.

Added link to view source code directly from
each sphinx documentation page.