Issue #162: Include and use module 'category'
All modules are assigned a numerical category code
within the comments at the beginning of the source
file. This is then used to organize the presentation
of the modules in the automatically generated sphinx
documentation. It is also used when generating the
dependency diagrams for the programs to simplify
them somewhat. This is done my ignoring the two
lowest level categories.

Also, updated the README to give some advice to run
the make_src_files.sh script whenever the dependencies
between the files has changed.