Issue #141: tt2phi coeff for tl/ad in unified subroutine.
In tt2phi_mod, the height coefficients for tangent linear/adjoint are
now module variables and are created in unified separate subroutine,
calcAltitudeCoef. The subroutine is called for the first time in
either tangent linear or adjoint to generate the coefficients. In the
following calls to tl/ad this subroutine skips execution.
  - Hence, the first parts of tl/ad responsible for the calculation
    of the coefficients are removed.

In tt2phi_mod, some of the constants from gps_mod are substituted
with the constants from mathphysconstants_mod.f90. The rest should
be either removed or relocated to another module (to be done later)

The logical variables to switch between the original and modified
GPSRO tangent linear/adjoint are removed. Now, the modified versions
are the only version used in the code.

Correction to 'zMT' variable is set similar to the master. zMT is not
used by the gps_struct1sw_v2 anymore, as the heights are generated
using tt2phi.

Remove print statements used for debugging.