Issue #80: Fix windRotation
The "fast" version of the wind vector rotation that
was implemented for LAM variational analysis on a rotated
analysis grid was being used also for the innovation
calculation. As a result it was used for interpolated
to observation located at the real equator, even though
the fast version blows up there (1/sin(lat)). Therefore,
a "nonlinear: nl" version was introduced for use in the
innovation calculation that follows exactly the calculation
in ezsint and the "fast" version is only used for the increment
when the analysis grid is rotated. This should be ok, since
we do not have any LAM configuration that include the real
equator.

Also, remove col_fillmvo and set DSTEPOBS in 2 system tests
back to 15 minutes for better comparison with the code in the
master and reduce wallclock for download tasks and some others.