Issue #141: Use static error estimation for GPSRO obs
Adding the gpsroDynError key to use static (or the default dynamic)
error estimation for the refractivity. This is for MIDAS development
for such as the impact of different height calculation, using the
GPSRO cost function as an estimation for omp. So the observation
error does not depend on the omp for GPSRO.

J.M. Aparicio recommended to ALWAYS set it to true (dynamic error)
for the operations.

Removing the knot -> m/s bug for the wind.