Issue #163: Ajustments for sea ice
The sea ice concentration from ice charts in SQLite files are in units
of per cent, while calculations for the assimilation are done in units
of concentration (ice fraction). So conversion from one to the other is
done. Note that we should avoid that type of conversion because it
makes the code more complicated. So in the future it would be good to
have the ice concentration as a fraction in the input files.