Issue #173: Second round of changes for VIS and WGE
The background check was adapted for the new observation types. Many missing changes from the first round of changes were also identified and fixed.

The code is now fully functionnal for assimilating visibility and wind gust.

However due to the highly non gaussian nature of visibility, the resulting analysis is currently poor. This problem will be addressed in the following commits.