Issue #173: Third round of changes for VIS and WGE
Due to the highly non gaussian nature of visibility, the code was modified to work in terms of the log of VIS for both the observations, the background and their associated errors. Therefore, the code now generates an analysis in terms of log(vis) that is then transformed in VIS in increment_mod.

In this version, observations are readed in terms of VIS. The transformation to log(VIS) is done in the observation operator.