Writing diagnostic files in sqlite format and modifying the data insertion into sqlite files
The content of obsSpaceData is stored into sqlite files.
The files are created on ramdisk. The ressources were adapted to it.
This commit contains also a modified verion of sqlr_insertSql routine
where the fortran missing values are replaced by 'null' sqlite values.
This makes the use of the sqlite files easier.