Issue #163: Use longer variable names.
As much as possible, the variable names should be more descriptive
and also longer than just one character so that doing a
search in the file for all uses of a variable is possible
(not the case for "i" and "j"). So I replaced "i" with "lonIndex"
and "j" with "latIndex".
This is what it is being used in most of the other modules.

Also include a blank space on either side of "=" and after ",".