Issue #163: Reading fields that have a mask.
In the case when typvar_in equal " "
(i.e. typvar is not used when selecting the record to read from the file),
then the attempt to read with the input typvar but with "@" in the second
position is not necessary. The logic is modified so that the attempt at
reading a masked field only happens when typvar_in is not equal to " ".

Otherwise the error message could say that reading the record with typvar equal to " @" failed. That could be confusing for users.