Issue #163: Memory allocation for the mask on processors.
The code was assuming that the mask is always allocated on processor 0,
but it is only allocated when the mask record is found in the
analysisgrid file. So now a broadcast of the logical variable
"maskAllocated = allocated(hco%mask)" is done from processor 0 to all
others. The information is then used to decide to allocate or not the
mask on a given processor.