Issue #163: Confusion between family and variable name.
The family name GL was used to also specify the analysis variable name.
This could be confusing. A cleaner and more flexible approach
is now to use the subroutine vnl_varNameFromVarNum so that it returns
"GL" when given as input BUFR_ICEC. This makes it more consistent with
other obs operators and makes explicit the link between observations
with element id BUFR_ICEC and the analysis variable name "GL". So a
call to "varName = vnl_varNameFromVarNum(BUFR_ICEC)" is added, and
"cdfam" is replaced with "varName" in the call to col_getElem.