Issue #186: scaleFactor for ensemble recentering
This commit addresses the issue of data extrapolation by the vertical
interpolation routine gsv_vInterpolate_r4.

The problem occurs in the context of ensemble recentering using
ensManip. Given a recentering mean having a higher model top than that
of the ensemble ensManip was generating large-amplitude, unrealistic
perturbations above the ensemble model top.

Two changes are implemented to 1) prevent accidental extrapolation
during vertical interpolation in any contexts and 2) enabling the use
of a scaling factor which forces the perturbations to zero on levels
where no information from the ensemble is available.

1- The vertical interpolation routine gsv_vInterpolate_r4 is modified
   so that by default it checks if the destination grid has a higher model
   top than the source grid. If this is the case, the program aborts.

2- This check can be skipped by providing a new "scalingFactor" parameter in
   ensManip's namelist. This user-defined scaling factor modulates the
   amplitude of the perturbations being applied to a given recentering mean.
   Setting this scaling factor to 0 for levels above the top of the source
   grid prevents the extrapolated perturbations from being applied to
   the recentering mean.