Issue #124: Approximate columng pressure using vgd_levels.
Instead of using the columng%vco to obtain the pressure on the
analysis levels, we do an approximation. Two hypothetical pressure
profiles (constructed based on 'columnhr P0' and
'columnhr and columng ip1' values) are used to compute the weights
to vertically interpolate the columnhr to columng.

Success 'run' listing for Var3D/gbc test on ppp2:
  - Feb 20, 12:51: original version (with GRTEST).

  - Feb 20, 17:09: when Psfc is from columnhr

  - Feb 20, 18:39: when Psfc is set to 1000 hPa.

Success 'run' listing for Var3D/splitobs_ppp test on ppp2:
  - Feb 20, 13:39: original version (with GRTEST).

  - Feb 20, 17:32: when Psfc is from columnhr

  - Feb 20, 18:40: when Psfc is set to 1000 hPa.