Issue #124: A new gsv_copy based on VarName used in gsv_readTrials.
The goal is to reduce the timing of the gsv_readTrials.

Since GZ/P are not read from the trials, copying these zero fields
and doing the horizontal/vertical interpolation in the gsv_readTrials
is expensive. A version of statevector, without GZ/P allocated, is
used to read the trials and do hintrp/vintrp to the analysis grid.
The gsv_copyByVarName is then used to copy the interpolated
statevector to the statevector_trials, which has GZ/P allocated.
The GZ/P calculation on the grid is performed in the next step.

Success 'run' listing for EnVar/gdps test on brooks:
  - Feb 27, 09:53: not reading wind from trial fields.

  - Feb 27, 16:06: use gsv_copyByVarName in vtr_setupTrials.