Issue #193: EnVar with lagged forecasts
The modifications to be able to compute Bens using lagged-forecast differences were made. A namelist variable (ensContainsFullField=.false.; default is .true.) must be added to specify when perturbations instead of full fields will be read by bmatrixEnsemble_mod.

The possibility to smooth the variances in Bens using a footprint operator was also added. The possibility to impose a domain-mean variances like in the actual Bnmc was also added.