Issue #185: Introduced 'OBS_CRPS' as new real body column in 'ObsSpaceData'
The EnKF has a separate list of variables in the obsspacedata. A new
real-body column is needed for the CRPS: Continuous Ranked Probability
Score. To facilitate the generation of summary scores, the CRPS will be
normalized with the standard deviation of the observational error. A
possible name for the new column is 'OBS_CRPS'.

The value of an ensemble is best measured with the CRPS. Having this
score readily available with facilitate verification and optimization of
the ensemble.