Issue #183: Fixed a type definition in 'sqlf_getDateStamp'
The interface of RMNLIB routine 'INCDATR' specifies that a 'real(8)'
must be given as last argument.  But in the 'obsIO' program, since
'OBS_REAL' type is at single precision ('real(4)'), the program 'obsIO'
aborted when compiled with '-check all'.  We impose a 'real(8)'
definition for variable 'delhh' and that fixed the problem.