Issue #193: gridBinning module
A new module to assign a bin category to each grid point was created. This allow to generate horizontal mean value based on e.g. the land/sea mask.

The footprint smoothing operator previously introduces was modified to be able to use the binning stategy information. It is now possible to smooth the data using only the grid points lying in the same bin category than the central point.