Issue #193: it is now possible to read LQ and LVIS perturbations
Modifications were made to allow the possibility to read variables that were previously tagged as "DIAG", for diagnostic variables, when working with perturbations.

The varType* in varnamelist_mod were removed. The variable to read (VIS if LVIS is wanted) and the data conversion is now made directly in gsv_readFile.