Issue #205: Modifications to be able to process oldstyle and FSR CrIS data in the same Midas run.
In the CrIS BURP files used now, the instrument code (BUFR element 2019)
is the same (620) for the two instruments.

The instrument code should be different for two different instruments
from Midas point of view.

A temporary modification was made to the code in order to modify the
value read forthat element from 620 to 2046 for CrIS FSR data
(based on the codtyp).

The definitive solution is to modify the BUFR decoding to set element
2019 to 2046 for CrIS FSR *or* to use information from BUFR element
002141 which is unfortunately not yet decoded.

The second solution is cleaner but may imply more work.