Issue #213: increment mask option is back
The option to mask the analysis increments in the blending zone (using an input mask field) was not previously transfer from addAnalInc to MIDAS. Now it is.

To activate this option, the user must add this to the namelist:

 &NAMINC
   useAnalIncMask=.true.
/

and provide a file named './analinc_mask' containing a MSKC 2D mask field.