Issue #224 Change default B HI matrix scalefactor from 1.0 to 0.0
This commit contains new code where the reading of namelist is done in the beginning of the B HI matrix setup.
And if the scalefactors of all the levels are equal to zero, the following computation is skipped.
This new behaviour needs the change in the namelists of the experiments that depend on the default values.