Issue #221: Reject observations for unknown satellites instead of aborting
This work aims at avoiding the program to abort when it reads unknown
satellites.  This happens sometimes when new satellites make their way
in the assimilation cycle but should not.  The configuration of the
MIDAS program are not ready to process those new satellites.
Unfortunately, the program was aborting.  Here, we just want to reject
the observations of the unknown satellites.