Issue #231: Limit the creation of database links for maestro test suite
By default, the tool 'make_links' from SSM domain 'cmdi/utils' that is
used in 'install_suite.sh' is creating the links as if the test suite
would run on all hosts available.  But, in practice, the suite is run
only on 'eccc-ppp1' and 'hare' or 'eccc-ppp2' and 'brooks'.  But, when
one of the hosts is not available (because of a failure or a system
time), the 'make_links' and so 'install_suite.sh' aborts because it
can't reach that host.  This is not such a problem for interactive use
but when the suite is created in a CI pipeline, the pipeline aborts.

With this commit, we create the links only for the hosts that are used
by the suite.  That means the value of 'FRONTEND' and 'BACKEND'
variables in the file 'resources/resources.def'.  We also add the
'TRUE_HOST' if one is using 'eccc-gpsc1' for example.  So, this is
reducing the probability of pipeline failure.

The list of hosts can be specified through a new environment variable:
      MIDAS_MAKE_LINKS_MACHINE_LIST

which will override the default behavior mentionned above.