Issue #231: Launch compiling job on same host as the test suite
The default machines to launch the compiling job were 'eccc-ppp1' and
'brooks'.  But this should be the same machines as the test suite is
using.  The script 'compile_all_plat.sh' has been modified to do so.

This new default behaviour can be override by setting the environment
variables 'COMPILING_MACHINE_PPP' and 'COMPILING_MACHINE_SUPER'.