Issue #124: Run GDPS with different MPI topology.
Changing the number of nodes from 27 to 36 (30) to test the timing in
GDPS. This requires changes the resource and running the splitobs
for the new MPI topology.

Add memory usage print for trapping the excess usage of proc 1. This
happened when the MPI topology changes from 12x27 to 12x36.

In oop_calcGPSGBJacobian if the calculated and original height differ
by 0.1 hPa, then print the error in the listing.

Write to listing first with name of s/b and then START or END.

Remove the timing outputs for calcPressure* and calcAltitudeCoeff*.