Issue #124: Bugfix for gsv_transposeTilesToVarsLevs of GZsfc.
For transposing the GZsfc from tiles to VarsLevs, the send/recv
arrays were allocated based on the dimension of statevector_out
(VarsLevs). The dimension of the send/recv arrays should be based on
the statevector_in (Tiles) in this case. This bug caused large
unnecessary spike in memory usage for transposing the 2D GZsfc field.

Running GDPS with 30 node.