Issue #124: Height/pressure alloc on the grid is true by default (major change)
Major changes to gridstatevector:
- In gridstatevector_mod, the default for the allocation of height/pressure
on the grid is set to .true. from now on. For height/pressure
allocation in gsv_allocate, in the presence of varNamrs_opt
Z_T/Z_M/P_T/P_M should be included in the varNamrs_opt. If
varNames_opt does not exist, DO NOT use allocHeight_opt=.true. and
allocPressure_opt=.true. as they are allocated by default. With this
configuration, there is only one way to allocate the height/pressure,
either with varNamrs_opt or no varNamrs_opt, at each time.

- Everywhere in the code where gsv_allocate is used without
varNames_opt, allocHeight_opt=.false. and allocPressure_opt=.false.
is forced to avoid unwanted allocation of height/pressure since the
default is true now. Hard-coded allocHeight_opt=.true. and
allocPressure_opt=.true. are removed since they are set only if they
are needed.

- The function checkColumnStateVectorMatch is added in
statetocolumn_mod. This function is called before s2c nl/tl/ad to
ensure the consistency of column and statevector.

- The vtr_transform for calculation of the height/pressure is called
only if the required variables are available in the statevector.

- In gsv_readFromFile, if the variable to set up the hco is not
found in the list then 'P0' is used as last option.

Major modification of columndata_mod:
- col_varExist accepts column optional argument now, similar to
gsv_varExist. The calls to the col_varExist in code are modified to
include the column.

- varExistList and nk are added to the columnData derived type (similar
to gridstatevector).

- Setting the varExistList of the height/pressure is moved from
col_setup to col_allocate. In col_allocate the vco is available,
numLevs are checked, and column heigh/pressure are allocated ONLY IF
needed. This is important for the 2D tests that do not need the
heigh/pressure.

- function 'col_getVarNameFromK' is added to enable to one-to-one
comparison of the column and statevector at each kIndex.

- change of the variable names for better understanding the content
(e.g numVar3D insteadof nvo3d)

For ensManip test, the varNames_opt is used as input to subroutines.
To ensure the height/pressure allocation is handled correctly, a
local variable varNames is used to copy the content of the
varNames_opt when varNames_opt is prescribed. Otherwise, varNames
includes the default varnames from the namelist. This is is necessary
to prevent allocation of height/pressure if they are not needed.

For ensembleH test, the trials need to be read in VarsLevs
distribution. A transpose is added to change the statevector to
Tiles distribution before calling the s2c_nl (s2c_nl is only called
in Tiles distribution).

Temporary bugfix in biascorrection_mod. The calculation of
statistics for the predictor fails when there is no obs.

Increase resources for var/vbc.