Issue #124: Allocate height/pressure if numLev>0 for column/statevector
Allocate height/pressure if numLev>0 in columndata and gridstatevector.

Check for existence of height/pressure before reading the column.

To calculate height/pressure on the grid, we only need to check the
existence of height/pressure in the statevector, instead of checking
for all the needed variables required for allocation of height/pressure.

Abort if one of the nlev_T/nlev_M is zero and the other one is
non-zero. We need a unified treatment of the thermodynamic and
momentum levels in the code (like in calcPressure, tt2phi, ...).

Change the input files for oMinusF/o-p_sst test so the nlev_T=nlev_M=1.