Issue #124: Changes before the merging to the master.
Using the vco%vcode instead of vco_get(vco%vgrid,...) in the code.

Nullify the pointers in the main subroutine to avoid initializing
them on the declaration line.

Use the varNames extracted from statevector for the allocation,
instead of using allocHeight_opt and allocPressure_opt, everywhere
in the code.

Removing the local variable addHeightSfcOffset in the subroutine
gsv_setup since there is a module variable with the same name.

Revising tt2phi_mod.f90 by removing the interface blocks and renaming
tt2phi_gsv* back to tt2phi*.

Moving gsv_allocate before the memberIndex loop in the ensembleH
program to avoid redundant allocation/deallocation in the loop.

Removing the unnecessary modules/bgcheck folder and
modules/calcpressure_mod.f90

Changes to the nitermax in namelist, and ncpus in resources to the
original values for two tests.

Changes to the formatting of the comments for some subroutines for
auto-documentation.

Removing some unnecessary comments in the code.

CHANGELOG.md is updated.