Issue #238: Allow changing lats/lons in the column to calculate interpolation weights.
Lats/lons are needed to calculate the interpolation weights.
Lats/lons are changed from one level to the next in the slanted
columns. The changing lats/lons with kIndex values are used to
calculate the interpolation weight and to allow transitioning to
slanted columns.

Increase in number of openMP threads for the following tests are
needed:
var/EnVar/24mem_xc40
var/EnVar/96mem_ppp_sqlite
var/vbc
var/bgck/conv
obsImpact/EFSO
obsImpact/HFSO

The check for the following tests fail and the results need update.
var/EnVar/24mem_xc40
var/EnVar/96mem_ppp_sqlite
var/vbc

We increased the resources for the above three tests in the master
by the same amount and the check fails. Further investigation
reveals that the original TOVS cost function value (before the
minimization) is affected by very small amount due to change in
number of threads in these three tests. This might be due to the
dependent openMP loops in TOVS library.